requirements.txt
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
idna==3.3
mypy-extensions==0.4.3
oauthlib==3.2.0
pathspec==0.9.0
platformdirs==2.5.1
python-dotenv==0.20.0
requests==2.27.1
requests-oauthlib==1.3.1
tomli==2.0.1
tweepy==4.8.0
typing_extensions==4.1.1
urllib3==1.26.9
